@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 2.4.0+28
 
 environment:
-  sdk: ">=2.16.2 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -33,19 +33,19 @@ dependencies:
   qr_flutter: ^4.0.0
   clipboard: ^0.1.3
   share_plus: ^4.0.4
-  url_launcher: ^6.1.0
+  url_launcher: ^6.1.2
   image_gallery_saver: ^1.7.1
   screenshot: ^1.2.3
-  shared_preferences: ^2.0.13
-  google_fonts: ^2.3.1
-  image_picker: ^0.8.5
+  shared_preferences: ^2.0.15
+  google_fonts: ^2.3.2
+  image_picker: ^0.8.5+3
   google_mlkit_barcode_scanning: ^0.0.2
-  camera: ^0.9.4+20
+  camera: ^0.9.4+24
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^1.0.4
+  flutter_lints: ^2.0.1
   flutter_launcher_icons: ^0.9.2
   flutter_native_splash: ^2.1.6
 
