@@ -4,14 +4,15 @@ import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import '../base_for_form/base_for_form.dart';
 
 class BodyFormTiktok extends BaseForm {
-  const BodyFormTiktok({Key? key}) : super(key: key);
+  const BodyFormTiktok({super.key});
 
   @override
-  _BodyFormTiktokState createState() => _BodyFormTiktokState();
+  State<BodyFormTiktok> createState() => _BodyFormTiktokState();
 }
 
 class _BodyFormTiktokState extends State<BodyFormTiktok> {
   final TextEditingController _textEditingController = TextEditingController();
+  late final Size _size;
 
   @override
   void dispose() {
@@ -19,15 +20,19 @@ class _BodyFormTiktokState extends State<BodyFormTiktok> {
     super.dispose();
   }
 
+  @override
+  void didChangeDependencies() {
+    _size = MediaQuery.of(context).size;
+    super.didChangeDependencies();
+  }
+
   String _filterToCreateQrcodeTiktok() =>
       _textEditingController.text.contains('https://www.tiktok.com/@')
           ? _textEditingController.text
-          : 'https://www.tiktok.com/@' + _textEditingController.text;
+          : 'https://www.tiktok.com/@${_textEditingController.text}';
 
   @override
   Widget build(BuildContext context) {
-    final Size _size = MediaQuery.of(context).size;
-
     return Form(
       key: widget.getKey,
       child: Column(
@@ -74,7 +79,9 @@ class _BodyFormTiktokState extends State<BodyFormTiktok> {
           Padding(
             padding: EdgeInsets.symmetric(horizontal: _size.width * 0.22),
             child: widget.makeButtoncreateQRCode(
-                context: context, filter: _filterToCreateQrcodeTiktok),
+                context: context,
+                filter: _filterToCreateQrcodeTiktok,
+                size: _size),
           )
         ],
       ),
