@@ -4,14 +4,14 @@ import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import '../../../../../shared/launch_link/launch_link.dart';
 
 class ButtonContact extends StatelessWidget with OpenLink {
-  const ButtonContact({Key? key}) : super(key: key);
+  const ButtonContact({super.key});
 
   @override
   Widget build(BuildContext context) {
-    final Size _size = MediaQuery.of(context).size;
+    final Size size = MediaQuery.of(context).size;
 
     return SizedBox(
-      height: _size.height * 0.09,
+      height: size.height * 0.09,
       child: ElevatedButton(
         style: ElevatedButton.styleFrom(
           elevation: 0,
@@ -20,7 +20,7 @@ class ButtonContact extends StatelessWidget with OpenLink {
         ),
         onPressed: () => openLink('mailto:ScannerQRCode@protonmail.com'),
         child: Padding(
-          padding: EdgeInsets.symmetric(horizontal: _size.width * 0.07),
+          padding: EdgeInsets.symmetric(horizontal: size.width * 0.07),
           child: Row(
             children: [
               Flexible(
