@@ -5,15 +5,14 @@ class ButtonHelpWidget extends StatelessWidget {
   final String solution;
 
   const ButtonHelpWidget(
-      {required this.problem, required this.solution, Key? key})
-      : super(key: key);
+      {required this.problem, required this.solution, super.key});
 
   @override
   Widget build(BuildContext context) {
-    final Size _size = MediaQuery.of(context).size;
+    final Size size = MediaQuery.of(context).size;
 
     return SizedBox(
-      height: _size.height * 0.09,
+      height: size.height * 0.09,
       child: ElevatedButton(
         style: ElevatedButton.styleFrom(
           elevation: 0,
@@ -26,7 +25,7 @@ class ButtonHelpWidget extends StatelessWidget {
               'solution': solution
             }),
         child: Padding(
-          padding: EdgeInsets.symmetric(horizontal: _size.width * 0.07),
+          padding: EdgeInsets.symmetric(horizontal: size.width * 0.07),
           child: Row(
             children: [
               const Spacer(),
